llm_quiz_generator/requirements.txt
google-generativeai
protobuf
grpcio
PyPDF2
python-dotenv
requests